+++ engine/runtime/src/main/java/org/enso/interpreter/node/expression/builtin/number/decimal/NegateNode.java
@@ -2,7 +2,7 @@ package org.enso.interpreter.node.expression.builtin.number.decimal;
 
 import org.enso.interpreter.dsl.BuiltinMethod;
 
-@BuiltinMethod(type = "Decimal", name = "negate", description = "Negation for numbers.")
+@BuiltinMethod(type = "Float", name = "negate", description = "Negation for numbers.")
 public class NegateNode extends FloatNode {
   double execute(double self) {
     return -self;
